refactor: flatten release version module
