+++ src/core/release/version.rs
@@ -1,6 +1,5 @@
 mod default_pattern_for_file;
 mod types;
-mod version;
 
 pub(crate) use default_pattern_for_file::{
     build_init_warnings, read_component_snapshot, resolve_version_file_path,
@@ -18,7 +17,6 @@ pub use types::{
     BumpResult, ChangelogValidationResult, ComponentVersionInfo, ComponentVersionSnapshot,
     UnconfiguredPattern, VersionTargetInfo,
 };
-pub use version::{get_component_version, increment_version, parse_version};
 
 use crate::core::component::{self, Component, VersionTarget};
 use crate::core::config::{from_str, set_json_pointer, to_string_pretty};
@@ -30,6 +28,54 @@ use crate::core::extension::ExtensionManifest;
 use crate::core::release::changelog;
 use serde_json::Value;
 use std::path::Path;
+
+/// Parse version from content using regex pattern.
+/// Pattern must contain a capture group for the version string.
+/// Content is trimmed to handle trailing newlines in VERSION files.
+pub fn parse_version(content: &str, pattern: &str) -> Option<String> {
+    text::extract_first(content, pattern)
+}
+
+/// Increment semver version or set an explicit version.
+///
+/// bump_type can be:
+/// - "patch", "minor", or "major" — increments the corresponding semver component
+/// - An explicit version string like "2.0.0" — returned as-is after validation
+pub fn increment_version(version: &str, bump_type: &str) -> Option<String> {
+    if bump_type.contains('.') {
+        let parts: Vec<&str> = bump_type.split('.').collect();
+        if parts.len() != 3 || parts.iter().any(|p| p.parse::<u32>().is_err()) {
+            return None;
+        }
+        return Some(bump_type.to_string());
+    }
+
+    let parts: Vec<&str> = version.split('.').collect();
+    if parts.len() != 3 {
+        return None;
+    }
+
+    let major: u32 = parts[0].parse().ok()?;
+    let minor: u32 = parts[1].parse().ok()?;
+    let patch: u32 = parts[2].parse().ok()?;
+
+    let (new_major, new_minor, new_patch) = match bump_type {
+        "patch" => (major, minor, patch + 1),
+        "minor" => (major, minor + 1, 0),
+        "major" => (major + 1, 0, 0),
+        _ => return None,
+    };
+
+    Some(format!("{}.{}.{}", new_major, new_minor, new_patch))
+}
+
+/// Get version string from a component's first version target.
+/// Returns None if no version targets configured or version can't be read.
+/// Use this for simple version checks (e.g., deploy outdated detection).
+pub fn get_component_version(component: &Component) -> Option<String> {
+    let target = component.version_targets.as_ref()?.first()?;
+    read_local_version(&component.local_path, target)
+}
 
 pub(crate) fn replace_versions(
     content: &str,
@@ -564,6 +610,53 @@ mod tests {
             result,
             " * @since 2.0.0\n * @since 1.0.0\n * @since 2.0.0\n"
         );
+    }
+
+    #[test]
+    fn increment_version_patch() {
+        assert_eq!(
+            increment_version("1.2.3", "patch"),
+            Some("1.2.4".to_string())
+        );
+    }
+
+    #[test]
+    fn increment_version_minor() {
+        assert_eq!(
+            increment_version("1.2.3", "minor"),
+            Some("1.3.0".to_string())
+        );
+    }
+
+    #[test]
+    fn increment_version_major() {
+        assert_eq!(
+            increment_version("1.2.3", "major"),
+            Some("2.0.0".to_string())
+        );
+    }
+
+    #[test]
+    fn increment_version_explicit_version() {
+        assert_eq!(
+            increment_version("1.25.5", "2.0.0"),
+            Some("2.0.0".to_string())
+        );
+        assert_eq!(
+            increment_version("0.5.0", "1.0.0"),
+            Some("1.0.0".to_string())
+        );
+    }
+
+    #[test]
+    fn increment_version_explicit_invalid() {
+        assert_eq!(increment_version("1.0.0", "2.0"), None);
+        assert_eq!(increment_version("1.0.0", "abc.def.ghi"), None);
+    }
+
+    #[test]
+    fn increment_version_unknown_bump_type() {
+        assert_eq!(increment_version("1.0.0", "huge"), None);
     }
 
     #[test]
